More Views Message
Final Message had 6 buttons. Now it has 4 plus 1 more if there are errors. A new button "More Views" shows the new message. It can show the In File as loaded into ST and Split 13,10. if this was end users I would not use Message() I would make a Window(). This is an example of how Nig Bang Theory can show ST info.